+++ src/components/Footer/Footer.stories.js
@@ -3,6 +3,7 @@ import { Footer } from "./Footer";
 import logo from "../../assets/wmms-blk.svg";
 import upArrow from "../../assets/upArrow.svg";
 import FR from "../../translations/fr.json";
+import EN from "../../translations/en.json";
 
 export default {
   title: "Components/Footer",
@@ -14,13 +15,72 @@ const Template = (args) => <Footer {...args} />;
 export const ErrorPageFooter = Template.bind({});
 ErrorPageFooter.args = {
   btnLink: "/",
-  lang: "error",
+  error: true,
 };
 
 export const EnFooter = Template.bind({});
 EnFooter.args = {
   btnLink: "/",
-  lang: "en",
+  landscapeLinks: [
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link1,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link2,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link3,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link4,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link5,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link6,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link7,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: FR.landscapeLinks.link8,
+    },
+    {
+      landscapeLink: "url to page",
+      landscapeLinkText: EN.landscapeLinks.link9,
+    },
+  ],
+  brandLinks: [
+    {
+      brandLink: "url to page",
+      brandLinkText: EN.brandLinks.link1,
+    },
+    {
+      brandLink: "url to page",
+      brandLinkText: EN.brandLinks.link2,
+    },
+    {
+      brandLink: "url to page",
+      brandLinkText: EN.brandLinks.link3,
+    },
+    {
+      brandLink: "url to page",
+      brandLinkText: EN.brandLinks.link4,
+    },
+    {
+      brandLink: "url to page",
+      brandLinkText: EN.brandLinks.link5,
+    },
+  ],
   logoUrl: logo,
   logoAltText: "Symbol of the Government of Canada",
   btnIcon: upArrow,
@@ -29,64 +89,63 @@ EnFooter.args = {
 export const FrFooter = Template.bind({});
 FrFooter.args = {
   btnLink: "/",
-  lang: "fr",
   landscapeLinks: [
     {
-      landscapeLink: FR.landscapeLinks.link1link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link1,
     },
     {
-      landscapeLink: FR.landscapeLinks.link2link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link2,
     },
     {
-      landscapeLink: FR.landscapeLinks.link3link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link3,
     },
     {
-      landscapeLink: FR.landscapeLinks.link4link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link4,
     },
     {
-      landscapeLink: FR.landscapeLinks.link5link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link5,
     },
     {
-      landscapeLink: FR.landscapeLinks.link6link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link6,
     },
     {
-      landscapeLink: FR.landscapeLinks.link7link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link7,
     },
     {
-      landscapeLink: FR.landscapeLinks.link8link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link8,
     },
     {
-      landscapeLink: FR.landscapeLinks.link9link,
+      landscapeLink: "url to page",
       landscapeLinkText: FR.landscapeLinks.link9,
     },
   ],
   brandLinks: [
     {
-      brandLink: FR.brandLinks.link1link,
+      brandLink: "url to page",
       brandLinkText: FR.brandLinks.link1,
     },
     {
-      brandLink: FR.brandLinks.link2link,
+      brandLink: "url to page",
       brandLinkText: FR.brandLinks.link2,
     },
     {
-      brandLink: FR.brandLinks.link3link,
+      brandLink: "url to page",
       brandLinkText: FR.brandLinks.link3,
     },
     {
-      brandLink: FR.brandLinks.link4link,
+      brandLink: "url to page",
       brandLinkText: FR.brandLinks.link4,
     },
     {
-      brandLink: FR.brandLinks.link5link,
+      brandLink: "url to page",
       brandLinkText: FR.brandLinks.link5,
     },
   ],